add /favorites delete functionality

+++ app/controllers/favorites_controller.rb
@@ -8,15 +8,19 @@ class FavoritesController < ApplicationController
     pet = Pet.find(params[:pet_id])
     favorite.add_pet(pet.id)
     session[:favorite] = favorite.pets
-    flash[:notice] = "You now have added #{pet.name} to your favorites!"
+    flash[:notice] = "You now have added this pet to your favorites!"
     redirect_to "/pets/#{pet.id}"
   end
 
   def destroy
-    pet = Pet.find(params[:pet_id])
-    favorite.delete_pet(pet.id)
-    session[:favorite] = favorite.pets
-    flash[:notice] = "You have deleted this Pet from your favorites!"
-    redirect_to "/pets/#{params[:pet_id]}"
+    if request.env['PATH_INFO'] == '/favorites'
+      favorite.remove_all
+    else
+      pet = Pet.find(params[:pet_id])
+      favorite.delete_pet(pet.id)
+      session[:favorite] = favorite.pets
+      flash[:notice] = "You have deleted this Pet from your favorites!"
+    end
+    redirect_to(request.env["HTTP_REFERER"])
   end
 end
